Use ProviderAccount with individual account retrieval
