+++ src/Components/LogLayout.tsx
@@ -1,9 +1,13 @@
 // react
-import {useCallback, useRef} from "react";
+import {useCallback, useEffect, useRef} from "react";
 // flex-layout
 import {Layout, Model, TabNode} from 'flexlayout-react';
 import 'flexlayout-react/style/light.css';
 import {Layout1} from '../Layouts/Layout1';
+// global vars
+import {useAccount} from "../Providers/ProviderAccount";
+// react router (get path)
+import {useLocation} from "react-router-dom";
 // split component
 import LogTable from "./LogTable"
 import LogColVisibility from "./LogColVisibility";
@@ -11,12 +15,14 @@ import LogTags from "./LogTags";
 import LogAccountSelector from "./LogAccountSelector";
 
 export default function LogLayout() {
+    // get path
+    const {pathname} = useLocation();
     // reference for flexlayout (used in the future)
     const layoutRef = useRef<Layout | null>(null);
-
+    // global vars
+    const {handleDefs} = useAccount();
     // flexlayout ui model
     const model:Model = Model.fromJson(Layout1)
-
     // renders components using specified flexlayout model
     const factory = useCallback((node:TabNode) => {
         // reads in model and retrieves values for "components"
@@ -29,6 +35,31 @@ export default function LogLayout() {
             case "Table": return <LogTable/>
         }
     }, [])
+
+    // get account data (via path/account)
+    // - returns empty arrays if account is not found
+    const fetchAccount = useCallback(async () => {
+        const res = await fetch("/api/getAccount",
+            {
+                method: "POST",
+                headers: {
+                    "Content-Type": "application/json"
+                },
+                body: JSON.stringify({name: pathname.substring(1)}),
+            });
+        if(!res.ok) {
+            console.error("Error occurred in fetchAccount: ", res.statusText);
+            return;
+        }
+        const data = await res.json();
+        handleDefs(data?.account?.AccName ?? [], data?.account?.RowData ?? [], data?.account?.ColDefs ?? [],
+            data?.account?.TagDefs ?? []);
+    }, [pathname, handleDefs])
+
+    useEffect(() => {
+        if(pathname === "/") return;
+        fetchAccount()
+    }, [pathname])
 
     return (
         <Layout
