+++ src/Components/LogTable.tsx
@@ -3,13 +3,13 @@ import { AgGridReact } from 'ag-grid-react';
 // ag grid: theme related
 import { themeAlpine } from "ag-grid-community";
 // global vars
-import { useGrid } from "../Providers/ProviderGrid";
+import {useAccount} from "../Providers/ProviderAccount";
 // styling
 import "./MUIStyling/LogTable.css";
 
 export default function LogTable() {
     // global vars
-    const { gridRef, rowData, colDefs } = useGrid();
+    const { gridRef, rowData, colDefs } = useAccount();
 
     return (
         <div className="GridContainer">
